feat(shared-modules-refactor): Sub-PR 1 — extraer <ProveedoresModule> + ADR-011
Auditoría 2026-04-27 detectó que el módulo Proveedores rompía la
convención que el resto del repo cumple: pages cross-empresa = wrapper
de ~30 líneas que delega a un componente shared. Antes de este PR,
`app/rdb/proveedores/page.tsx` (1579 líneas) y
`app/dilesa/proveedores/page.tsx` (1578 líneas) estaban duplicados al
100% (1535 líneas idénticas literal, 23 de diff cosmético).

## Cambios

Componente nuevo:
- `components/proveedores/proveedores-module.tsx` — todo el flujo de
  Proveedores parametrizado por props (empresaId, empresaSlug, logoPath,
  membreteAlt). Lista, alta con CSF, edit, archivar/restaurar, dedup
  por RFC, update CSF con diff selectivo. Sin cambios funcionales.

Pages reducidos:
- `app/rdb/proveedores/page.tsx`: 1579 → 17 líneas.
- `app/dilesa/proveedores/page.tsx`: 1578 → 17 líneas.

Constantes centralizadas:
- `lib/empresa-constants.ts` (nuevo) — single source of truth de UUIDs
  (RDB + DILESA por ahora; ANSA/COAGAN cuando entren al repo).
- `lib/dilesa-constants.ts` re-exporta `DILESA_EMPRESA_ID` desde el
  archivo nuevo y queda marcado como legacy para no romper los 30+
  call sites existentes.

ADR-011 (Sub-PR 2 incluido en mismo PR):
- `docs/adr/011_shared_modules_cross_empresa.md` codifica la convención
  con 5 reglas (SM1-SM5): page como wrapper de delegación, componente
  en `components/<modulo>/`, props parametrizan diferencias, cuándo NO
  aplica, code review enforza.

## Bug encontrado durante extracción

El modal de RFC duplicado en DILESA decía hardcoded "ya está registrado
como proveedor activo en RDB" (copy-paste sin actualizar de RDB).
Resuelto en el componente shared con `empresaSlug.toUpperCase()` como
label dinámico — exactamente el drift accidental que SM5 quiere evitar.

## Cierre de iniciativa

Estado `planned → in_progress`. Próximo hito: Sub-PR 3
`juntas-detail-shared` (Tier A, ~1346 líneas duplicadas).

## Validación

- typecheck ✅
- lint ✅ (1 warning heredado del page original sobre `<img>`,
  pre-existente; CI lo tolera)
- format:check ✅
- 222 tests ✅

Smoke pendiente en preview deploy:
- /rdb/proveedores: lista, alta nueva con CSF, edit, archivar.
- /dilesa/proveedores: idem + verificar logo/membrete DILESA.
